Rename subgraph module to avoid conflict
