+++ networkx/classes/subgraphviews.py
