+++ networkx/classes/tests/test_subgraphviews.py
